requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
requests
langchain
langchain-openai
langchain-community
faiss-cpu
openai
tiktoken
pypdf